Limit websockets message sizes to ints

+++ src/rddms_io/client.py
@@ -1717,7 +1717,7 @@ class rddms_connect:
         data_partition_id: str | None = None,
         authorization: str | SecretStr | None = None,
         etp_timeout: float | None = None,
-        max_message_size: float = 2**20,
+        max_message_size: int = 2**20,
         use_compression: bool = True,
     ) -> None:
         self.uri = uri
